Split content script updater in separate file.
