+++ src/background/background.js
@@ -1,3 +1,4 @@
 import './background-setup.js'
 import './open-tutorial.js'
+import './update-content.js'
 import '../store/store-background.js'
